+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/profiler.php
@@ -7,7 +7,6 @@ $container->loadFromExtension('framework', [
     'php_errors' => ['log' => true],
     'profiler' => [
         'enabled' => true,
-        'collect_serializer_data' => true,
     ],
     'serializer' => [
         'enabled' => true,